new version

+++ Novice/main.cpp
@@ -1,5 +1,5 @@
 #include <Novice.h>
-#include <math.h>
+#include <cmath>
 
 const char kWindowTitle[] = "GC2A_02_テイン_タイ_アウン";
 
@@ -25,30 +25,30 @@ Matrix4x4 Multiply(Matrix4x4 m1, Matrix4x4 m2) {
 Matrix4x4 MakeRotateXMatrix(float radius) {
 	Matrix4x4 result = {};
 	result.m[0][0] = 1.0f;
-	result.m[1][1] = cosf(radius);
-	result.m[1][2] = sinf(radius);
-	result.m[2][1] = -sinf(radius);
-	result.m[2][2] = cosf(radius);
+	result.m[1][1] = std::cos(radius);
+	result.m[1][2] = std::sin(radius);
+	result.m[2][1] = -std::sin(radius);
+	result.m[2][2] = std::cos(radius);
 	result.m[3][3] = 1.0f;
 	return result;
 }
 Matrix4x4 MakeRotateYMatrix(float radius) {
 	Matrix4x4 result = {};
-	result.m[0][0] = cosf(radius);
-	result.m[0][2] = -sinf(radius);
+	result.m[0][0] = std::cos(radius);
+	result.m[0][2] = -std::sin(radius);
 	result.m[1][1] = 1.0f;
-	result.m[2][0] = sinf(radius);
-	result.m[2][2] = cosf(radius);
+	result.m[2][0] = std::sin(radius);
+	result.m[2][2] = std::cos(radius);
 	result.m[3][3] = 1.0f;
 	return result;
 }
 Matrix4x4 MakeRotateZMatrix(float radius) {
 
 	Matrix4x4 result = {};
-	result.m[0][0] = cosf(radius);
-	result.m[0][1] = sinf(radius);
-	result.m[1][0] = -sinf(radius);
-	result.m[1][1] = cosf(radius);
+	result.m[0][0] = std::cos(radius);
+	result.m[0][1] = std::sin(radius);
+	result.m[1][0] = -std::sin(radius);
+	result.m[1][1] = std::cos(radius);
 	result.m[2][2] = 1.0f;
 	result.m[3][3] = 1.0f;
 	return result;
